WIP: use ObjectProperty and ObjectMethod instead of Property
