+++ test/lib/elements/types/ExportDefaultDeclaration.js
@@ -44,7 +44,7 @@ describe('ExportDefaultDeclaration', () => {
     });
 
     it('should accept a named class', () => {
-        var statement = parseAndGetStatement('export default class C{};');
+        var statement = parseAndGetStatement('export default class C {};');
         expect(statement.declaration.type).to.equal('ClassDeclaration');
     });
