+++ test/lib/elements/types/JSXAttribute.js
@@ -12,7 +12,7 @@ describe('JSXAttribute', () => {
         expect(expression.openingElement.attributes[0].value).to.equal(null);
     });
 
-    it('allow self closing tag containing an attribute with a literal', () => {
+    it('allow self closing tag containing an attribute with a StringLiteral', () => {
         const expression = parseAndGetExpression('< a b = "3" / >');
         expect(expression.openingElement.attributes.length).to.equal(1);
         expect(expression.openingElement.attributes[0].type).to.equal('JSXAttribute');
@@ -23,7 +23,7 @@ describe('JSXAttribute', () => {
         expect(expression.openingElement.attributes[0].value.raw).to.equal('"3"');
     });
 
-    it('allow self closing tag containing an namespaced attribute with a literal', () => {
+    it('allow self closing tag containing an namespaced attribute with a StringLiteral', () => {
         const expression = parseAndGetExpression('< a b:c = "3" / >');
         expect(expression.openingElement.attributes.length).to.equal(1);
         expect(expression.openingElement.attributes[0].type).to.equal('JSXAttribute');
