+++ test/lib/elements/types/SpreadProperty.js
@@ -25,7 +25,7 @@ describe('SpreadProperty', () => {
 
     it('should allow other params', () => {
         let properties = parseAndGetStatement('let a = { b, ... c, ... d } ;').declarations[0].init.properties;
-        expect(properties[0].type).to.equal('Property');
+        expect(properties[0].type).to.equal('ObjectProperty');
         expect(properties[0].key.type).to.equal('Identifier');
         expect(properties[0].key.name).to.equal('b');
         expect(properties[0].value.type).to.equal('Identifier');
